Use int for units.gridUnit
It was an int already anyway, so let's be true about this and also
return an int. This reflects pixel-alignment in the API, not just in the
implementation.

+++ src/declarativeimports/core/units.cpp
@@ -96,7 +96,7 @@ void Units::setDevicePixelRatio(const qreal scale)
     }
 }
 
-qreal Units::gridUnit() const
+int Units::gridUnit() const
 {
     qDebug() << "FontMetrics: " << QApplication::font().pixelSize() << QFontMetrics(QApplication::font()).boundingRect("M");
     qDebug() << " MRect" << QFontMetrics(QApplication::font()).boundingRect("M").size();
